VL53L5CX-BNO08X-viewer-main/VL53L5CX-BNO08X-viewer-main/viewer/requirements.txt
viser>=1.0.0
pyserial>=3.5
numpy>=1.20.0
scipy>=1.10.0
trimesh>=4.0.0
Pillow>=9.0.0
